+++ Changes.md
@@ -17,6 +17,7 @@ API
 - MetadataAlgo :
   - Added optional `user` argument to `addAnnotationTemplate()`.
   - Added optional `userOnly` argument to `annotationTemplates()`.
+- AnnotationsGadget : Added `setVisibleAnnotations()` and `getVisibleAnnotations()` methods to allow filtering of annotations.
 
 Breaking Changes
 ----------------